% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpeaks.control.R
\name{findpeaks.control}
\alias{findpeaks.control}
\title{Control parameters for ZIMHMM and ZIHMM}
\usage{
findpeaks.control(epsilon.em = c(0.001, 0.001, 0.001, 0.001),
  epsilon.inner.em = 0.001, maxit.em = 500, minit.em = 3,
  gap.em = 3, maxcount.em = 3, max.phi = 1000, min.sigma2 = 1e-08,
  max.sigma2 = 10, maxcount.inner.em = 50, criterion = "MULTI",
  min.zero = .Machine$double.xmin, pcut = 0.05, quiet = F)
}
\arguments{
\item{epsilon.em}{Either a positive value or a vector of size 4 with the convergence tolerance values for the EM algorithm (see 'criterion' below). Default is c(1e-3,1e-3,1e-3,1e-3)}

\item{epsilon.inner.em}{positive convergence tolerance for the conditional maximizations used in ZIMHMM}

\item{maxit.em}{integer giving the maximum number of EM iterations (default 500)}

\item{minit.em}{integer giving the minimum number of EM iterations to start evaluating the convergence (default 3)}

\item{gap.em}{integer giving the number of EM iterations apart to compute the convergence criterion (default 3)}

\item{maxcount.em}{integer giving the number of consecutive EM iterations satisfying the convergence criterion  in order to stop the algorithm (default 3)}

\item{max.phi}{maximum positive value allowed for the dispersion parameters (default 1000)}

\item{min.sigma2}{minimum positive value allowed for the variance component (default 1e-08)}

\item{max.sigma2}{maximum positive value allowed for the variance component (default 10)}

\item{maxcount.inner.em}{integer giving the maxium number of conditional maximizations used in ZIMHMM (default 50)}

\item{criterion}{convergence criterion: either "MRCPE" (maximum absolute relative change in parameter estimates), "MACPE" (maximum absolute change of parameter estimates),
"ARCEL" (absolute relative change of the Q-function), "ACC" (agreement of Viterbi peak calls),
or "MULTI" (simultaneously check for MRCPE, MACPE, ARCEL, and ACC).
For ACC, it computes the percentage of windows falling in the main diagonal of a 2 by 2 table of Viterbi predictions 'gap.em' iterations apart. Default is "MULTI"}

\item{min.zero}{minimum positive value allowed in computations to avoid having zeros (default is .Machine$double.xmin)}

\item{pcut}{cutoff for rejection controlled EM algorithm (default 0.05)}

\item{quiet}{whether to print messages (default F)}
}
\value{
A list with components equal to the arguments
}
\description{
This function passes controlling parameters for ZIMHMM and ZIHMM. Most of these parameters control the EM algorithm
}
\examples{
data(H3K36me3.Huvec)
ChIP = as.matrix(H3K36me3.Huvec[,c("H3K36me3.Huvec.Rep1","H3K36me3.Huvec.Rep2","H3K36me3.Huvec.Rep3")])
Control = log(as.matrix(H3K36me3.Huvec[,c("Control.Huvec.Rep1","Control.Huvec.Rep2","Control.Huvec.Rep3")])+1)
offset = matrix(0,nrow = nrow(ChIP),ncol = ncol(ChIP))
control = findpeaks.control(maxit.em = 100)
ZIMHMM(ChIP = ChIP,Control = Control,offset = offset,random = 'intercept',control = control)

}
\references{
\url{https://github.com/plbaldoni/ZIMHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
