% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMM.R
\name{HMM}
\alias{HMM}
\title{Hidden Markov Model (HMM)}
\usage{
HMM(ChIP.init, Control.init, offset.init, pcut = 0.05,
  epsilon.em = 0.001, maxit.em = 5, minit.em = 3, gap.em = 3,
  maxcount.em = 3, max.phi = 1000, min.zero = .Machine$double.xmin,
  quant = 0.75, quiet = F)
}
\arguments{
\item{ChIP.init}{M*N matrix of ChIP read counts, where M is the number of windows in the analyzed genome and N is the number of replicates}

\item{Control.init}{M*N matrix of log-transformed Control read counts}

\item{offset.init}{M*N matrix of offsets. If no offset is used, use offset = matrix(0,nrow=M,ncol=N)}

\item{pcut}{cutoff for rejection controlled EM algorithm (default is 0.05)}

\item{epsilon.em}{A positive value with the convergence tolerance value for the EM algorithm (default is 1e-3)}

\item{maxit.em}{integer giving the maximum number of EM iterations (default 5)}

\item{minit.em}{integer giving the minimum number of EM iterations to start evaluating the convergence (default 3)}

\item{gap.em}{integer giving the number of EM iterations apart to compute the convergence criterion (default 3)}

\item{maxcount.em}{integer giving the number of consecutive EM iterations satisfying the convergence criterion  in order to stop the algorithm (default 3)}

\item{max.phi}{maximum positive value allowed for the dispersion parameters (default 1000)}

\item{min.zero}{minimum positive value allowed in computations to avoid having zeros (default is .Machine$double.xmin)}

\item{quant}{quantile of the distribution of the score of scaled ChIP counts to define the very first set of enrichment and background windows to begin the EM algorithm (default is 0.75).
Scaled windows with score below the quant-tile are defined as background to start the EM algorithm.}

\item{quiet}{whether to print messages (default F)}
}
\value{
A list with components
\item{Pi}{Vector of initial probabilities of the HMM}
\item{Gamma}{Matrix of transition probabilities of the HMM}
\item{Psi}{Vector of component-specific parameters of the HMM}
\item{Prob}{Mx2 Matrix with posterior probabilities}
\item{LogF}{Mx2 Matrix with log-forward probabilities}
\item{LogB}{Mx2 Matrix with log-backward probabilities}
\item{Loglik}{Mx2 Matrix with window-based probabilities}
\item{Parhist}{Matrix with paramater estimates across EM iterations}
\item{Mean}{M*(N*2) Matrix with NB means for every replicate and HMM component. The first two columns of Mean are the background and enrichment means of replicate 1, respectively, and so on}
\item{Viterbi}{Predicted sequence of Viterbi states}
}
\description{
This function runs a 2-state HMM with negative binomial emissions. It is used for parameter initialization in ZIMHMM and ZIHMM.
}
\examples{
data(Huvec)
ChIP = SummarizedExperiment::assay(Huvec,'ChIP')
Control = log(SummarizedExperiment::assay(Huvec,'Control')+1)
offset = matrix(0,nrow = nrow(ChIP),ncol = ncol(ChIP),byrow = TRUE)
\dontrun{HMM(ChIP.init = rowSums(ChIP),Control.init = rowMeans(Control),offset.init = rowMeans(offset))}

}
\references{
\url{https://github.com/plbaldoni/ZIMHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
