% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBed.R
\name{makeBed}
\alias{makeBed}
\title{Create a BED file from an output of ZIMHMM}
\usage{
makeBed(output, data, file, tracklabel = "User Track")
}
\arguments{
\item{output}{An output from `ZIMHMM()`}

\item{data}{A RangedSummarizedExperiment object with the same assay used in `ZIMHMM()`}

\item{file}{The full path and name of the output file (e.g. './example.bed')}

\item{tracklabel}{Optional label for the track (default is 'User Track')}
}
\value{
A GRanges object with consensus peaks.
}
\description{
`makeBed()` imports the output from `ZIMHMM()` and a RangedSummarizedExperiment object dataset to
create a BED file to be used in downstream analyses.
}
\examples{
data(Huvec)
ChIP = SummarizedExperiment::assay(Huvec,'ChIP')
Control = log(SummarizedExperiment::assay(Huvec,'Control')+1)
offset = matrix(0,nrow = nrow(ChIP),ncol = ncol(ChIP))
\dontrun{output = ZIMHMM(ChIP = ChIP,Control = Control,offset = offset,random = 'intercept',control = controlPeaks())}
\dontrun{makeBed(output,data = Huvec,file = './H3K36me3.bed')}

}
\references{
\url{https://github.com/plbaldoni/mixHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
