% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIHMM.R
\name{ZIHMM}
\alias{ZIHMM}
\title{Zero Inflated Fixed Effects Hidden Markov Model (ZIHMM)}
\usage{
ZIHMM(ChIP, Control, offset, control)
}
\arguments{
\item{ChIP}{M*N matrix of ChIP read counts, where M is the number of windows in the analyzed genome and N is the number of replicates}

\item{Control}{M*N matrix of log-transformed Control read counts}

\item{offset}{M*N matrix of offsets (use offset = matrix(0,nrow=M,ncol=N) for no offset)}

\item{control}{list of control arguments (use control = findpeaks.control() for default controls)}
}
\value{
None
}
\description{
This function runs ZIHMM (under developement).
}
\examples{
ZIHMM()

}
\references{
\url{https://github.com/plbaldoni/ZIMHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
