% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIMHMM.R
\name{ZIMHMM}
\alias{ZIMHMM}
\title{Zero Inflated Mixed Effects Hidden Markov Model (ZIMHMM)}
\usage{
ZIMHMM(ChIP, Control, offset, random, control)
}
\arguments{
\item{ChIP}{M*N matrix of ChIP read counts, where M is the number of windows in the analyzed genome and N is the number of replicates}

\item{Control}{M*N matrix of log-transformed Control read counts}

\item{offset}{M*N matrix of offsets (use offset = matrix(0,nrow=M,ncol=N) for no offset)}

\item{random}{either 'intercept' for random intercept model or 'slope' for random slope model}

\item{control}{list of control arguments (use control = findpeaks.control() for default controls)}
}
\value{
None
}
\description{
This function runs ZIMHMM (under developement).
}
\examples{
ZIMHMM()

}
\references{
\url{https://github.com/plbaldoni/ZIMHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
