% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOffset.R
\name{createOffset}
\alias{createOffset}
\title{Create Offsets for Peak Calling}
\usage{
createOffset(ChIP, method = "sum", span = 0.3)
}
\arguments{
\item{ChIP}{M*N matrix of ChIP read counts, where M is the number of windows in the analyzed genome and N is the number of experiments and replicates}

\item{method}{either total sum of read counts (method="sum"), smoothed MA trend (method="loess"), or median ratio (method="ratio").}

\item{span}{proportion of data to be used in the smoothing (see loessFit from limma, default 0.3)}
}
\value{
M*N matrix of offsets
}
\description{
This function create offsets for peak calling. The current implementation can output three different set of offsets:
total sum of read counts (method="sum"), smoothed MA trend (method="loess", similar to csaw), or median ratio (method="ratio", similar to DESeq2).
}
\examples{
data(Huvec)
ChIP = SummarizedExperiment::assay(Huvec,'ChIP')
offset = createOffset(ChIP,method="loess")

}
\references{
\url{https://github.com/plbaldoni/ZIMHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
