% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeaks.R
\name{plotPeaks}
\alias{plotPeaks}
\title{Create a plot of ChIP read counts with called peaks and posterior probabilities}
\usage{
plotPeaks(output, chr, ranges, data, peaks = NULL)
}
\arguments{
\item{output}{Either an output from `ZIMHMM()` or `ZIHMM()`}

\item{chr}{A string with the chromosome name to be displayed (such as 'chr19')}

\item{ranges}{A vector with two elements that specifies the genomic windows to be plotted.}

\item{data}{A RangedSummarizedExperiment object with the same assay (with the `ChIP` counts) used in `ZIMHMM()`}

\item{peaks}{An optional vector of length M with zeros (background) and ones (peaks) representing the enriched windows. If not provided,
`plotPeaks()` will use the Viterbi sequence from `output`.}
}
\value{
A ggplot
}
\description{
`plotPeaks()` imports the output from `ZIMHMM()` or `ZIHMM()` and plot the peak
calls of a given genomic region.
}
\examples{
data(Huvec)
ChIP = SummarizedExperiment::assay(Huvec,'ChIP')
Control = log(SummarizedExperiment::assay(Huvec,'Control')+1)
offset = matrix(log(colSums(ChIP)),nrow = nrow(ChIP),ncol = ncol(ChIP),byrow = TRUE)
\dontrun{output = ZIHMM(ChIP = ChIP,Control = Control,offset = offset,control = controlPeaks(epsilon.em = 1e-3,criterion = 'MRCPE'))}
\dontrun{plotPeaks(output = output,chr = 'chr19',ranges = c(16898093,17040483),data = Huvec)}

}
\references{
\url{https://github.com/plbaldoni/ZIMHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
