% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIHMM.R
\name{ZIHMM}
\alias{ZIHMM}
\title{Zero Inflated Fixed Effects Hidden Markov Model (ZIHMM)}
\usage{
ZIHMM(ChIP, Control, offset, control)
}
\arguments{
\item{ChIP}{M*N matrix of ChIP read counts, where M is the number of windows in the analyzed genome and N is the number of replicates}

\item{Control}{M*N matrix of log-transformed Control read counts}

\item{offset}{M*N matrix of offsets. If no offset is used, use offset = matrix(0,nrow=M,ncol=N)}

\item{control}{list of control arguments from controlPeaks()}
}
\value{
A list with components
\item{Pi}{Vector of initial probabilities of the HMM}
\item{Gamma}{Matrix of transition probabilities of the HMM}
\item{Psi}{Vector of component-specific parameters of the HMM}
\item{Zeroinfl}{M*N Matrix with zero-inflation probabilities}
\item{Prob}{Mx2 Matrix with posterior probabilities}
\item{LogF}{Mx2 Matrix with log-forward probabilities}
\item{LogB}{Mx2 Matrix with log-backward probabilities}
\item{Loglik}{Mx2 Matrix with window-based probabilities}
\item{Parhist}{Matrix with paramater estimates across EM iterations}
\item{Mean}{M*(N*2) Matrix with NB means for every replicate and HMM component. The first two columns of Mean are the background and enrichment means of replicate 1, respectively, and so on}
\item{Viterbi}{Predicted sequence of Viterbi states}
}
\description{
This function runs ZIHMM.
}
\examples{
data(H3K36me3.Huvec)
ChIP = as.matrix(H3K36me3.Huvec[,c("H3K36me3.Huvec.Rep1","H3K36me3.Huvec.Rep2","H3K36me3.Huvec.Rep3")])
Control = log(as.matrix(H3K36me3.Huvec[,c("Control.Huvec.Rep1","Control.Huvec.Rep2","Control.Huvec.Rep3")])+1)
offset = matrix(log(colSums(ChIP)),nrow = nrow(ChIP),ncol = ncol(ChIP),byrow = TRUE)
\dontrun{ZIHMM(ChIP = ChIP,Control = Control,offset = offset,control = controlPeaks(epsilon.em = 1e-3,criterion = 'MRCPE'))}

}
\references{
\url{https://github.com/plbaldoni/ZIMHMM}
}
\author{
Pedro L. Baldoni, \email{pedrobaldoni@gmail.com}
}
